Agregada la detección de caída libre e imágenes del coyote

+++ TP2/Proyectos/merge_definitivo/source/display_utils.h
@@ -23,7 +23,7 @@ extern "C" {
 
 void print_image(uint8_t x0, uint8_t y0, unsigned char* image, uint8_t width_in_px, uint8_t height_in_px);
 void print_falling_state(void);
-void print_accel_display_state(float acceleration);
+void print_accel_display_state(int acceleration);
 
 /*==================[cplusplus]==============================================*/
 #ifdef __cplusplus
